Moved User.cs from Database to Models for better organization
